# Use function `babbageMinUTxOValue` for faster minimum UTxO calculations.