## Issue Number

ADP-2144

## Summary

This PR adjusts the `{compute,isBelow}MinimumCoinForUTxO` functions so that in the Babbage era only, we bypass the Cardano API function `calculateMinimumUTxO`, and instead directly use the ledger function `babbageMinUTxOValue`.

This leads to a significant reduction in latency when computing transaction fees:

**Before:**
```
    Latencies for 2 fixture wallets scenario
        postTransactionFee  - 3177.0 ms
    Latencies for 10 fixture wallets scenario
        postTransactionFee  - 3555.3 ms
```

**After:**
```
    Latencies for 2 fixture wallets scenario
        postTransactionFee  - 13.5 ms
    Latencies for 10 fixture wallets scenario
        postTransactionFee  - 15.6 ms
```