@jonathanknowles 2022-08-19
> Triggered nightly latency step: https://buildkite.com/input-output-hk/cardano-wallet-nightly/builds/1467#_

Results:
```
    Latencies for 2 fixture wallets scenario
        postTransactionFee  - 12.5 ms
    Latencies for 10 fixture wallets scenario
        postTransactionFee  - 13.0 ms
    Latencies for 100 fixture wallets
        postTransactionFee  - 23.1 ms
    Latencies for 2 fixture wallets with 10 txs scenario
        postTransactionFee  - 13.3 ms
    Latencies for 2 fixture wallets with 20 txs scenario
        postTransactionFee  - 14.5 ms
    Latencies for 2 fixture wallets with 100 txs scenario
        postTransactionFee  - 16.6 ms
    Latencies for 10 fixture wallets with 10 txs scenario
        postTransactionFee  - 13.9 ms
    Latencies for 10 fixture wallets with 20 txs scenario
        postTransactionFee  - 15.2 ms
    Latencies for 10 fixture wallets with 100 txs scenario
        postTransactionFee  - 18.2 ms
    Latencies for 2 fixture wallets with 100 utxos scenario
        postTransactionFee  - 16.4 ms
    Latencies for 2 fixture wallets with 200 utxos scenario
        postTransactionFee  - 20.6 ms
    Latencies for 2 fixture wallets with 500 utxos scenario
        postTransactionFee  - 24.5 ms
    Latencies for 2 fixture wallets with 1000 utxos scenario
        postTransactionFee  - 46.9 ms
```