@Anviking 2022-08-19
And results from recent master for comparison:
```
   Latencies for 2 fixture wallets scenario
        postTransactionFee  - 2851.2 ms
    Latencies for 10 fixture wallets scenario
        postTransactionFee  - 3049.6 ms
    Latencies for 100 fixture wallets
        postTransactionFee  - 11185.9 ms
    [this run seem to have errored for the rest of the results]
```