@jonathanknowles 2022-08-19
> I'm kind of astonished how we'd be able to see this kind of extreme slowdown from the cardano-api wrapper... and only the babbage era thinking might laziness be involved somehow?

It might very well be! 

I did run some experiments that added strictness annotations in certain places, but none of these experiments were fruitful. More analysis is needed: I'm very keen to work out exactly what the root cause is!